Create Tax Reporting tool.py
This code is a Python application that automates the process of reporting tax details for corporate actions. The app allows the user to open two Excel files, one containing client information and the other containing cash slips. The user can then enter ISIN and AP Tunnus values, and the app will write this information, along with the data from the two Excel files, to a destination Excel file. The app uses the Tkinter library for the user interface, and the Pandas and Openpyxl libraries for reading and writing the Excel files. The user can open the files via the "Client List" and "Cash Slip" buttons and once data is entered in the respective fields, the user can click on the "Report" button to write the data to the destination Excel file. The app will then save the changes to the destination file.